requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Dev
black>=24.0.0
ruff>=0.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
